+++ bukkit/changelog.md
@@ -1,3 +1,4 @@
 - Try to fall back to spigot mode if version is unknown
 - Fixed inconsistency with server implementation compared to mods
 - Improved invalid packet handling on servers
+- Improved paper version parsing